third_party/blink/web_tests/external/wpt/css/css-fonts/idlharness-expected.txt

This is a testharness.js-based test.
Found 57 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSSFontFaceDescriptors interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface object length
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface object name
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute src
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontFamily
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-family
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontStyle
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-style
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontWeight
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-weight
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontStretch
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-stretch
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontWidth
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-width
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontSize
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-size
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute sizeAdjust
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute size-adjust
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute unicodeRange
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute unicode-range
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontFeatureSettings
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-feature-settings
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontVariationSettings
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-variation-settings
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontNamedInstance
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-named-instance
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontDisplay
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-display
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute fontLanguageOverride
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute font-language-override
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute ascentOverride
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute ascent-override
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute descentOverride
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute descent-override
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute lineGapOverride
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute line-gap-override
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute superscriptPositionOverride
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute superscript-position-override
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute subscriptPositionOverride
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute subscript-position-override
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute superscriptSizeOverride
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute superscript-size-override
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute subscriptSizeOverride
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceDescriptors interface: attribute subscript-size-override
  assert_own_property: self does not have own property "CSSFontFaceDescriptors" expected property "CSSFontFaceDescriptors" missing
[FAIL] CSSFontFaceRule interface: attribute style
  assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
[FAIL] CSSFontFeatureValuesRule interface: attribute historicalForms
  assert_true: The prototype object must have a property "historicalForms" expected true got false
[FAIL] CSSFontFeatureValuesMap interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSFontFeatureValuesMap" expected property "CSSFontFeatureValuesMap" missing
[FAIL] CSSFontFeatureValuesMap interface object length
  assert_own_property: self does not have own property "CSSFontFeatureValuesMap" expected property "CSSFontFeatureValuesMap" missing
[FAIL] CSSFontFeatureValuesMap interface object name
  assert_own_property: self does not have own property "CSSFontFeatureValuesMap" expected property "CSSFontFeatureValuesMap" missing
[FAIL] CSSFontFeatureValuesMap interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSFontFeatureValuesMap" expected property "CSSFontFeatureValuesMap" missing
[FAIL] CSSFontFeatureValuesMap interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSFontFeatureValuesMap" expected property "CSSFontFeatureValuesMap" missing
[FAIL] CSSFontFeatureValuesMap interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSFontFeatureValuesMap" expected property "CSSFontFeatureValuesMap" missing
[FAIL] CSSFontFeatureValuesMap interface: maplike<CSSOMString, [object Object]>
  Cannot read properties of undefined (reading 'prototype')
[FAIL] CSSFontFeatureValuesMap interface: operation set(CSSOMString, (unsigned long or sequence<unsigned long>))
  assert_own_property: self does not have own property "CSSFontFeatureValuesMap" expected property "CSSFontFeatureValuesMap" missing
Harness: the test ran to completion.
